fix: Fix broken ref on TextInput
 - Merge TextInput components back
- Split render methods
