--- src/components/TextInput/TextInputFlat.js
@@ -1,478 +1,0 @@
-/* @flow */
-
-import * as React from 'react';
-import {
-  View,
-  Animated,
-  TextInput as NativeTextInput,
-  StyleSheet,
-  I18nManager,
-} from 'react-native';
-import { polyfill } from 'react-lifecycles-compat';
-import color from 'color';
-import Text from '../Typography/Text';
-import type { Props, State, RenderProps } from './types';
-
-const AnimatedText = Animated.createAnimatedComponent(Text);
-
-const MINIMIZED_LABEL_Y_OFFSET = -12;
-const MAXIMIZED_LABEL_FONT_SIZE = 16;
-const MINIMIZED_LABEL_FONT_SIZE = 12;
-const LABEL_WIGGLE_X_OFFSET = 4;
-const FOCUS_ANIMATION_DURATION = 150;
-const BLUR_ANIMATION_DURATION = 180;
-const LABEL_PADDING_HORIZONTAL = 12;
-const RANDOM_VALUE_TO_CENTER_LABEL = 4; // Don't know why 4, but it works
-
-class TextInputFlat extends React.Component<Props, State> {
-  static defaultProps = {
-    disabled: false,
-    error: false,
-    multiline: false,
-    editable: true,
-    render: (props: RenderProps) => <NativeTextInput {...props} />,
-  };
-
-  static getDerivedStateFromProps(nextProps: Props, prevState: State) {
-    return {
-      value:
-        typeof nextProps.value !== 'undefined'
-          ? nextProps.value
-          : prevState.value,
-    };
-  }
-
-  state = {
-    labeled: new Animated.Value(this.props.value || this.props.error ? 0 : 1),
-    error: new Animated.Value(this.props.error ? 1 : 0),
-    focused: false,
-    placeholder: this.props.error ? this.props.placeholder : '',
-    value: this.props.value,
-    labelLayout: {
-      measured: false,
-      width: 0,
-    },
-  };
-
-  componentDidUpdate(prevProps: Props, prevState: State) {
-    if (
-      prevState.focused !== this.state.focused ||
-      prevState.value !== this.state.value ||
-      prevProps.error !== this.props.error
-    ) {
-      // The label should be minimized if the text input is focused, or has text
-      // In minimized mode, the label moves up and becomes small
-      if (this.state.value || this.state.focused || this.props.error) {
-        this._minmizeLabel();
-      } else {
-        this._restoreLabel();
-      }
-    }
-
-    if (
-      prevState.focused !== this.state.focused ||
-      prevProps.label !== this.props.label ||
-      prevProps.error !== this.props.error
-    ) {
-      // Show placeholder text only if the input is focused, or has error, or there's no label
-      // We don't show placeholder if there's a label because the label acts as placeholder
-      // When focused, the label moves up, so we can show a placeholder
-      if (this.state.focused || this.props.error || !this.props.label) {
-        this._showPlaceholder();
-      } else {
-        this._hidePlaceholder();
-      }
-    }
-
-    if (prevProps.error !== this.props.error) {
-      // When the input has an error, we wiggle the label and apply error styles
-      if (this.props.error) {
-        this._showError();
-      } else {
-        this._hideError();
-      }
-    }
-  }
-
-  componentWillUnmount() {
-    clearTimeout(this._timer);
-  }
-
-  _showPlaceholder = () => {
-    clearTimeout(this._timer);
-
-    // Set the placeholder in a delay to offset the label animation
-    // If we show it immediately, they'll overlap and look ugly
-    this._timer = setTimeout(
-      () =>
-        this.setState({
-          placeholder: this.props.placeholder,
-        }),
-      50
-    );
-  };
-
-  _hidePlaceholder = () =>
-    this.setState({
-      placeholder: '',
-    });
-
-  _timer: TimeoutID;
-  _root: ?NativeTextInput;
-
-  _showError = () => {
-    Animated.timing(this.state.error, {
-      toValue: 1,
-      duration: FOCUS_ANIMATION_DURATION,
-      useNativeDriver: true,
-    }).start(this._showPlaceholder);
-  };
-
-  _hideError = () => {
-    Animated.timing(this.state.error, {
-      toValue: 0,
-      duration: BLUR_ANIMATION_DURATION,
-      useNativeDriver: true,
-    }).start();
-  };
-
-  _restoreLabel = () =>
-    Animated.timing(this.state.labeled, {
-      toValue: 1,
-      duration: FOCUS_ANIMATION_DURATION,
-      useNativeDriver: true,
-    }).start();
-
-  _minmizeLabel = () =>
-    Animated.timing(this.state.labeled, {
-      toValue: 0,
-      duration: BLUR_ANIMATION_DURATION,
-      useNativeDriver: true,
-    }).start();
-
-  _handleFocus = (...args) => {
-    if (this.props.disabled) {
-      return;
-    }
-
-    this.setState({ focused: true });
-
-    if (this.props.onFocus) {
-      this.props.onFocus(...args);
-    }
-  };
-
-  _handleBlur = (...args) => {
-    if (this.props.disabled) {
-      return;
-    }
-
-    this.setState({ focused: false });
-
-    if (this.props.onBlur) {
-      this.props.onBlur(...args);
-    }
-  };
-
-  _handleChangeText = (value: string) => {
-    if (!this.props.editable) {
-      return;
-    }
-
-    this.setState({ value });
-    this.props.onChangeText && this.props.onChangeText(value);
-  };
-
-  /**
-   * @internal
-   */
-  setNativeProps(...args: any) {
-    return this._root && this._root.setNativeProps(...args);
-  }
-
-  /**
-   * Returns `true` if the input is currently focused, `false` otherwise.
-   */
-  isFocused() {
-    return this._root && this._root.isFocused();
-  }
-
-  /**
-   * Removes all text from the TextInput.
-   */
-  clear() {
-    return this._root && this._root.clear();
-  }
-
-  /**
-   * Focuses the input.
-   */
-  focus() {
-    return this._root && this._root.focus();
-  }
-
-  /**
-   * Removes focus from the input.
-   */
-  blur() {
-    return this._root && this._root.blur();
-  }
-
-  render() {
-    const {
-      disabled,
-      label,
-      error,
-      selectionColor,
-      underlineColor,
-      style,
-      theme,
-      render,
-      multiline,
-      ...rest
-    } = this.props;
-
-    const { colors, fonts } = theme;
-    const fontFamily = fonts.regular;
-    const hasActiveOutline = this.state.focused || error;
-
-    let inputTextColor, activeColor, underlineColorCustom, placeholderColor;
-
-    if (disabled) {
-      inputTextColor = activeColor = color(colors.text)
-        .alpha(0.54)
-        .rgb()
-        .string();
-      placeholderColor = colors.disabled;
-      underlineColorCustom = 'transparent';
-    } else {
-      inputTextColor = colors.text;
-      activeColor = error ? colors.error : colors.primary;
-      placeholderColor = colors.placeholder;
-      underlineColorCustom = underlineColor || colors.disabled;
-    }
-
-    const containerStyle = {
-      backgroundColor: theme.dark
-        ? color(colors.background)
-            .lighten(0.24)
-            .rgb()
-            .string()
-        : color(colors.background)
-            .darken(0.06)
-            .rgb()
-            .string(),
-      borderTopLeftRadius: theme.roundness,
-      borderTopRightRadius: theme.roundness,
-    };
-
-    const labelHalfWidth = this.state.labelLayout.width / 2;
-    const baseLabelTranslateX =
-      (I18nManager.isRTL ? 1 : -1) *
-      (1 - MINIMIZED_LABEL_FONT_SIZE / MAXIMIZED_LABEL_FONT_SIZE) *
-      labelHalfWidth;
-
-    const labelStyle = {
-      fontFamily,
-      fontSize: MAXIMIZED_LABEL_FONT_SIZE,
-      transform: [
-        {
-          // Wiggle the label when there's an error
-          translateX: this.state.error.interpolate({
-            inputRange: [0, 0.5, 1],
-            outputRange: [
-              0,
-              this.state.value && error ? LABEL_WIGGLE_X_OFFSET : 0,
-              0,
-            ],
-          }),
-        },
-        {
-          // Move label to top
-          translateY: this.state.labeled.interpolate({
-            inputRange: [0, 1],
-            outputRange: [MINIMIZED_LABEL_Y_OFFSET, 0],
-          }),
-        },
-        {
-          // Make label smaller
-          scale: this.state.labeled.interpolate({
-            inputRange: [0, 1],
-            outputRange: [
-              MINIMIZED_LABEL_FONT_SIZE / MAXIMIZED_LABEL_FONT_SIZE,
-              1,
-            ],
-          }),
-        },
-        {
-          // Offset label scale since RN doesn't support transform origin
-          translateX: this.state.labeled.interpolate({
-            inputRange: [0, 1],
-            outputRange: [
-              baseLabelTranslateX > 0
-                ? baseLabelTranslateX +
-                  labelHalfWidth / LABEL_PADDING_HORIZONTAL -
-                  RANDOM_VALUE_TO_CENTER_LABEL
-                : baseLabelTranslateX -
-                  labelHalfWidth / LABEL_PADDING_HORIZONTAL +
-                  RANDOM_VALUE_TO_CENTER_LABEL,
-              0,
-            ],
-          }),
-        },
-      ],
-    };
-
-    return (
-      <View style={[containerStyle, style]}>
-        <Animated.View
-          style={[
-            styles.underline,
-            {
-              backgroundColor: error
-                ? colors.error
-                : this.state.focused
-                  ? activeColor
-                  : underlineColorCustom,
-              // Underlines is thinner when input is not focused
-              transform: [{ scaleY: this.state.focused ? 1 : 0.5 }],
-            },
-          ]}
-        />
-
-        {label ? (
-          // Position colored placeholder and gray placeholder on top of each other and crossfade them
-          // This gives the effect of animating the color, but allows us to use native driver
-          <View
-            pointerEvents="none"
-            style={[
-              StyleSheet.absoluteFill,
-              {
-                opacity:
-                  // Hide the label in minimized state until we measure it's width
-                  this.state.value || this.state.focused
-                    ? this.state.labelLayout.measured
-                      ? 1
-                      : 0
-                    : 1,
-              },
-            ]}
-          >
-            <AnimatedText
-              onLayout={e =>
-                this.setState({
-                  labelLayout: {
-                    width: e.nativeEvent.layout.width,
-                    measured: true,
-                  },
-                })
-              }
-              style={[
-                styles.placeholder,
-                styles.placeholderFlat,
-                labelStyle,
-                {
-                  color: activeColor,
-                  opacity: this.state.labeled.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [hasActiveOutline ? 1 : 0, 0],
-                  }),
-                },
-              ]}
-              numberOfLines={1}
-            >
-              {label}
-            </AnimatedText>
-            <AnimatedText
-              style={[
-                styles.placeholder,
-                styles.placeholderFlat,
-                labelStyle,
-                {
-                  color: placeholderColor,
-                  opacity: hasActiveOutline ? this.state.labeled : 1,
-                },
-              ]}
-              numberOfLines={1}
-            >
-              {label}
-            </AnimatedText>
-          </View>
-        ) : null}
-
-        {render(
-          ({
-            ...rest,
-            ref: c => {
-              this._root = c;
-            },
-            onChangeText: this._handleChangeText,
-            placeholder: label
-              ? this.state.placeholder
-              : this.props.placeholder,
-            placeholderTextColor: placeholderColor,
-            editable: !disabled,
-            selectionColor:
-              typeof selectionColor === 'undefined'
-                ? activeColor
-                : selectionColor,
-            onFocus: this._handleFocus,
-            onBlur: this._handleBlur,
-            underlineColorAndroid: 'transparent',
-            multiline,
-            style: [
-              styles.input,
-              this.props.label
-                ? styles.inputFlatWithLabel
-                : styles.inputFlatWithoutLabel,
-              {
-                color: inputTextColor,
-                fontFamily,
-                textAlignVertical: multiline ? 'top' : 'center',
-              },
-            ],
-          }: RenderProps)
-        )}
-      </View>
-    );
-  }
-}
-
-polyfill(TextInputFlat);
-
-export default TextInputFlat;
-
-const styles = StyleSheet.create({
-  placeholder: {
-    position: 'absolute',
-    left: 0,
-    fontSize: 16,
-    paddingHorizontal: LABEL_PADDING_HORIZONTAL,
-  },
-  placeholderFlat: {
-    top: 19,
-  },
-  underline: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    bottom: 0,
-    height: 2,
-  },
-  input: {
-    flexGrow: 1,
-    paddingHorizontal: 12,
-    fontSize: 16,
-    margin: 0,
-    minHeight: 58,
-    textAlign: I18nManager.isRTL ? 'right' : 'left',
-    zIndex: 1,
-  },
-  inputFlatWithLabel: {
-    paddingTop: 24,
-    paddingBottom: 6,
-  },
-  inputFlatWithoutLabel: {
-    paddingVertical: 15,
-  },
-});
