+++ src/index.js
@@ -44,7 +44,7 @@ export { default as Surface } from './components/Surface';
 export { default as Switch } from './components/Switch';
 export { default as Appbar } from './components/Appbar/Appbar';
 export { default as TouchableRipple } from './components/TouchableRipple';
-export { default as TextInput } from './components/TextInput/TextInput';
+export { default as TextInput } from './components/TextInput';
 export {
   default as ToggleButton,
 } from './components/ToggleButton/ToggleButton';
